Move Turbine tests to flowsheet_control_test module
