+++ foqus_lib/unit_test/node_test.py
@@ -25,7 +25,6 @@ from foqus_lib.framework.graph.graph import Graph
 from foqus_lib.framework.graph.nodeModelTypes import nodeModelTypes
 from foqus_lib.framework.pymodel.pymodel import pymodel
 from foqus_lib.framework.pymodel import pymodel_test
-from foqus_lib.framework.sim.turbineConfiguration import TurbineConfiguration
 
 from importlib import import_module
 from pathlib import Path
@@ -874,57 +873,6 @@ class TestNode:
                 assert getattr(node.gr.output[node.name][vkey], attribute) == getattr(
                     v, attribute
                 )
-
-    def test_setSim_modelTurbine(self, node):
-        # manually add turbine model to test
-
-        # examples/test_files/Optimization/Model_Files/
-        # BFB_sinter_config_v6.2.json and BFB_cost_v6.2.3.json
-
-        turbdir = os.path.join(
-            os.path.abspath(__file__),
-            "../../../examples/tutorial_files/SimSinter/Tutorial_2",
-        )
-        turbpath = os.path.join(turbdir, "Flash_Example_AP.json")
-
-        # create config block and upload model files to Turbine
-        node.gr.turbConfig = TurbineConfiguration()
-        node.gr.turbConfig.writeConfig(overwrite=True)
-        node.gr.turbConfig.uploadSimulation(
-            simName="Flash_Example",
-            sinterConfigPath=os.path.normpath(turbpath),
-            update=True,
-            otherResources=[],
-        )
-
-        # set simulation
-        node.setSim(newModel="Flash_Example", newType=2)
-
-    def test_runTurbineCalc(self, node):
-        # manually add turbine model to test
-
-        # examples/test_files/Optimization/Model_Files/
-        # BFB_sinter_config_v6.2.json and BFB_cost_v6.2.3.json
-
-        turbdir = os.path.join(
-            os.path.abspath(__file__),
-            "../../../examples/tutorial_files/SimSinter/Tutorial_2",
-        )
-        turbpath = os.path.join(turbdir, "Flash_Example_AP.json")
-
-        # create config block and upload model files to Turbine
-        node.gr.turbConfig = TurbineConfiguration()
-        node.gr.turbConfig.writeConfig(overwrite=True)
-        node.gr.turbConfig.uploadSimulation(
-            simName="Flash_Example",
-            sinterConfigPath=os.path.normpath(turbpath),
-            update=True,
-            otherResources=[],
-        )
-
-        # set simulation
-        node.setSim(newModel="Flash_Example", newType=2)
-        node.runCalc  # covers node.runTurbineCalc
 
     def test_setSim_modelMLAI_example1(self, node, model_files):
         # skip this test if tensorflow is not available
